+++ CHANGELOG.md
@@ -10,6 +10,11 @@ New:
 * Add PeriodicTransformXD and VectorPeriodicTransformXD functions to support the data simulated with periodic boundary conditions. (#387)
 * Add CylindricalTransform and VectorCylindricalTransform to transform functions from cylindrical to Cartesian coordinates. (#387)
 * Add the kind attribute to RayTransferPipelineXD that determines whether the ray transfer matrix is multiplied by sensitivity ('power') or not ('radiance'). (#412)
+* Add focal_length property to the Beam. (#419)
+
+Bug fixes:
+* Fix beam width and density evaluation.
+* Fix beam direction evaluation. (#414)
 
 
 Release 1.4.0 (3 Feb 2023)